# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Veni AI Sustainability Cockpit API",
    "version": "v1"
  },
  "paths": {
...
    "/integrations/connectors": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Create Or Update Integration Connector",
        "operationId": "create_or_update_integration_connector_integrations_connectors_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationConfigCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/sync": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Sync Integrations",
        "operationId": "sync_integrations_integrations_sync_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/sync-jobs/{job_id}": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Get Sync Job",
        "operationId": "get_sync_job_integrations_sync_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Tenant Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorSyncJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/facts": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Project Facts",
        "operationId": "list_project_facts_projects__project_id__facts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Tenant Id"
            }
          },
          {
            "name": "metric_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric Code"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectFactsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectorSyncJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "integration_config_id": {
            "type": "string",
            "title": "Integration Config Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "connector_type": {
            "type": "string",
            "title": "Connector Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "inserted_count": {
            "type": "integer",
            "title": "Inserted Count"
          },
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "cursor_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Before"
          },
          "cursor_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor After"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At Utc"
          },
          "completed_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At Utc"
          },
          "diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostics"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "integration_config_id",
          "tenant_id",
          "project_id",
          "connector_type",
          "status",
          "current_stage",
          "record_count",
          "inserted_count",
          "updated_count"
        ],
        "title": "ConnectorSyncJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegrationConfigCreateRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "minLength": 1,
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "minLength": 1,
            "title": "Project Id"
          },
          "connector_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Connector Type"
          },
          "display_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Display Name"
          },
          "auth_mode": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Auth Mode",
            "default": "configured"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "resource_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Path"
          },
          "mapping_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Mapping Version",
            "default": "v1"
          },
          "connection_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Connection Payload"
          },
          "sample_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sample Payload"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "project_id",
          "connector_type",
          "display_name"
        ],
        "title": "IntegrationConfigCreateRequest"
      },
      "IntegrationConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "connector_type": {
            "type": "string",
            "title": "Connector Type"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "auth_mode": {
            "type": "string",
            "title": "Auth Mode"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "resource_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Path"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mapping_version": {
            "type": "string",
            "title": "Mapping Version"
          },
          "last_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Cursor"
          },
          "last_synced_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At Utc"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "connector_type",
          "display_name",
          "auth_mode",
          "status",
          "mapping_version"
        ],
        "title": "IntegrationConfigResponse"
      },
...
      "IntegrationSyncRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "minLength": 1,
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "minLength": 1,
            "title": "Project Id"
          },
          "connector_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Connector Ids"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "project_id"
        ],
        "title": "IntegrationSyncRequest"
      },
      "IntegrationSyncResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/ConnectorSyncJobResponse"
            },
            "type": "array",
            "title": "Jobs"
          },
          "synced_connector_count": {
            "type": "integer",
            "title": "Synced Connector Count"
          }
        },
        "type": "object",
        "required": [
          "jobs",
          "synced_connector_count"
        ],
        "title": "IntegrationSyncResponse"
      },
...
      "ProjectFactResponse": {
        "properties": {
          "fact_id": {
            "type": "string",
            "title": "Fact Id"
          },
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "period_key": {
            "type": "string",
            "title": "Period Key"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_numeric": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Numeric"
          },
          "value_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Text"
          },
          "source_system": {
            "type": "string",
            "title": "Source System"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "freshness_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness At Utc"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Ref"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "fact_id",
          "metric_code",
          "metric_name",
          "period_key",
          "source_system",
          "source_record_id"
        ],
        "title": "ProjectFactResponse"
      },
      "ProjectFactsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectFactResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ProjectFactsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}